% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casing.R
\name{casing}
\alias{casing}
\title{Finds values in a vector directly above and below a number}
\usage{
casing(x, into)
}
\arguments{
\item{x}{a number}

\item{into}{a vector where to find the values directly above and
below x}
}
\value{
a vector of the values of "into" vector directly above and below x
respectively
}
\description{
Finds values in a vector directly above and below a number
}
\examples{
casing(0.21,c(0.3,0.4,0.1,0.2))

}
\seealso{
Similar function : \code{\link{encase}}
}
