% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{collections}
\alias{collections}
\alias{oufti99}
\title{Collections of symbols}
\description{
\describe{
  \item{\strong{oufti99}}{First experimental symbol dataset, developed in
  Liege University by Anne-Christine Da Silva, Michiel Arts and Sébastien
  Wouters}
}
}
\examples{
\dontrun{
plot_collection(oufti99, name = "Oufti99")}

}
