% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Converts x values having an index into n values defined by the same
y index}
\usage{
convert(x, xindex, n, nindex)
}
\arguments{
\item{x}{a vector}

\item{xindex}{the index for each x value (vector of same length
than x)}

\item{n}{a vector of the values into which to convert the x values}

\item{nindex}{the index for each n value (vector of same length
than n)}
}
\description{
Converts x values having an index (of y values for instance)
into n values defined by the same index (but having possibly more values)
}
\examples{
x      <- c(10,20)
xindex <- c(1,2)

n      <- seq(0.1,1,by = 0.1)
nindex <- 1:length(n)

convert(x,xindex,n,nindex)

}
