% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earinc.R
\name{earinc}
\alias{earinc}
\title{Recalculates inclination in equal area projection}
\usage{
earinc(inc)
}
\arguments{
\item{inc}{inclination of the data; it is the angle from the
horizontal, is positive downward, and ranges from +90° for straight down to
-90° for straight up (Tauxe, 2010).}
}
\description{
Recalculates the inclination in equal area projection
}
\examples{
earinc(20)

}
