% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encase.R
\name{encase}
\alias{encase}
\title{Encases two numbers between multiples of a given number}
\usage{
encase(x1, x2, n)
}
\arguments{
\item{x1}{the first value of the interval}

\item{x2}{the second value of the interval (can be higher or lower,
but never equal to x1)}

\item{n}{the number to find the multiples from}
}
\value{
the multiples of n directly encompassing x1 and x2
}
\description{
Encases two numbers between multiples of a given number
}
\examples{
encase(5,1,5)

}
\seealso{
Similar function : \code{\link{casing}}
}
