% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encircle.R
\name{encircle}
\alias{encircle}
\title{Draws circles}
\usage{
encircle(
  r = 1,
  x = 0,
  y = 0,
  ndiv = 360,
  plot = TRUE,
  add = TRUE,
  output = FALSE,
  ...
)
}
\arguments{
\item{r}{the radius of the circles (of length 1 or n)}

\item{x}{the x value of the centre of the circles (of length 1 or n)}

\item{y}{the y value of the centre of the circles (of length 1 or n)}

\item{ndiv}{the number of segments making the circles}

\item{plot}{whether to plot the circles}

\item{add}{whether to add to an existing plot}

\item{output}{whether to return an output}

\item{...}{graphical parameters to feed to lines}
}
\value{
a list of x and y matrices having n rows, one for each circle
}
\description{
Draws circles
}
\examples{
plot(0, 0, xlim = c(-1,1), ylim = c(-1,1), asp = 1)

encircle(lwd = 2)
encircle(r = seq(0.1,0.9,0.1))

}
