% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastPairs.R
\name{fastPairs}
\alias{fastPairs}
\title{Scatterplot Matrices for large dataset}
\usage{
fastPairs(dat, bins = 60, base_size = 12, cor_size = 6)
}
\arguments{
\item{dat}{data frame of all parameters to be compared}

\item{bins}{numeric vector giving number of bins in both vertical and
horizontal directions. Set to 60 by default.}

\item{base_size}{size of the base texts (axis, titles, ...)}

\item{cor_size}{size of the text for the correlation coefficient. Has
different units (?) than base_size}
}
\description{
Makes a matrix of plots with a given data set of large
size using hexagonal binning (similar result than the pairs function)
}
\examples{
fastPairs(iris[1:4],bins=10)
}
