% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder.R
\name{folder}
\alias{folder}
\title{Creates a new folder where wanted if it does not exist yet}
\usage{
folder(dir, name)
}
\arguments{
\item{dir}{directory containing the folder}

\item{name}{name of the folder}
}
\value{
the directory of the folder itself
}
\description{
Creates a new folder where wanted if it does not exist yet
}
\examples{
# # To run example uncomment all: put in a script, select all and use
# #  ctrl+shift+c
#
# folder(getwd(),"test")

}
