% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formFunction.R
\name{formFunction}
\alias{formFunction}
\title{Converts a formula into a function}
\usage{
formFunction(formula)
}
\arguments{
\item{formula}{the formula to be converted. Should be of the form y ~ f(x)}
}
\description{
Converts a formula into a function
}
\examples{
f <- formFunction(y ~ log10(x))

f(x=1:10)
}
