% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat.lag.R
\name{mat.lag}
\alias{mat.lag}
\alias{mat.lead}
\title{Find the "next" or "previous" values in a matrix.}
\usage{
mat.lag(m, n = 1L, default = NA)

mat.lead(m, n = 1L, default = NA)
}
\arguments{
\item{m}{a matrix of values}

\item{n}{a positive integer of length 1, giving the number of
positions to lead or lag by}

\item{default}{value used for non-existent rows. Defaults to NA.}
}
\description{
Find the "next" or "previous" values in a matrix.
}
\examples{
m <- matrix(1:120, ncol = 12)

mat.lag(m)
mat.lead(m)

}
