% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.lim.R
\name{mid.lim}
\alias{mid.lim}
\title{Provides mid-points intervals in an ordered vector}
\usage{
mid.lim(x, id = 1L, b = "[]")
}
\arguments{
\item{x}{an ordered vector}

\item{id}{a vector of n interval IDs (default is 1 for each interval)}

\item{b}{a character vector for the interval boundaries rules, see
\code{as.lim} help page for details}
}
\value{
a lim object of intervals with boundaries at midway between the x
values
}
\description{
Provides mid-points intervals in an ordered vector
}
\examples{
mid.lim(c(1,3,7,20,45,63))

}
\seealso{
\code{\link{as.lim}}
}
