% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minorAxisTicks.R
\name{minorAxisTicks}
\alias{minorAxisTicks}
\title{Compute Pretty Minor Axis Tick Scales}
\usage{
minorAxisTicks(usr, n = NULL, at.maj = NULL, extend = T)
}
\arguments{
\item{usr}{the user coordinates of the minimum and maximum limits of the axis}

\item{n}{the number of intervals defined by the minor ticks}

\item{at.maj}{the positions at which major tick-marks are to be drawn.
By default (when NULL) tickmark locations are computed buy the axisTicks
function}

\item{extend}{whether to add minor ticks even outside the major ticks
(T) or not (F)}
}
\description{
Compute pretty mark locations for minor ticks, based on the way
that traditional R graphics do it.
}
\examples{
minorAxisTicks(usr = c(-20, 620), n = 10)

}
\seealso{
\code{\link{minorAxis}}, \code{\link{seq_log}}

This function is based on \code{\link{every_nth}}, which suppresses values
every multiple of a given number.
}
