% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfind.R
\name{pkgfind}
\alias{pkgfind}
\title{Find a specific pattern in the code of functions in a package}
\usage{
pkgfind(pkg, pattern)
}
\arguments{
\item{pkg}{a character string of the package to search in}

\item{pattern}{the pattern to search in the codes of the functions in the
package}
}
\value{
a vector of the names of the functions in which the
pattern is identified
}
\description{
This function names all the functions in a package that contain
a specific character pattern, typically the name of a function.
}
\examples{
pkgfind("StratigrapheR", "every_nth")

}
