% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_mult.R
\name{seq_mult}
\alias{seq_mult}
\title{Sequence ordered by multiple}
\usage{
seq_mult(l, mult, inv = FALSE)
}
\arguments{
\item{l}{the length of the sequence, or an object convertile into a vector
from which to determine the length from}

\item{mult}{the multiple to order by}

\item{inv}{whether to change mult into l/mult}
}
\description{
Sequence ordered by multiple
}
\examples{
seq_mult(10, 2)

seq_mult(15,3)

seq_mult(24,8)

seq_mult(seq(0.5,12,0.5),8)

seq_mult(10,2)[seq_mult(10,5)]

seq_mult(10,2)[seq_mult(10,2, inv = TRUE)]

}
