% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat.mean.R
\name{strat.mean}
\alias{strat.mean}
\title{Extrapolate and intrapolate tie points}
\usage{
strat.mean(dt, events = NULL, sections = NULL)
}
\arguments{
\item{dt}{a matrix of depth (or time) of the different tie points. Columns
are for the sections, rows for each tie point}

\item{events}{the name of the tie points}

\item{sections}{the name of the sections}
}
\description{
Extrapolate and intrapolate of stratigraphical tie points or
events, based on their position in different sections
}
\examples{
dt     <- tie.points.example[,2:6]
events <- tie.points.example[,1]

strat.mean(dt = dt, events = events)

}
