% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting.examples.R
\name{StratigrapheR.examples}
\alias{StratigrapheR.examples}
\alias{bed.example}
\alias{boundary.example}
\alias{chron.example}
\alias{example.ammonite}
\alias{example.ammonite.svg}
\alias{example.belemnite}
\alias{example.breccia}
\alias{example.lense}
\alias{example.liquefaction}
\alias{fossil.example}
\alias{log.loop.tex}
\alias{proxy.example}
\alias{proxy.example.litho}
\alias{zeq_example}
\title{Data for examples}
\description{
Supporting data sets to use in the examples. They will be used in the
examples. example.ammonite.svg and log.loop.tex are meant to generate their
respective .svg and .tex files.
}
\details{
\describe{
  \item{\strong{Litholog drawing data}}{bed.example, boundary.example,
  example.ammonite, example.ammonite.svg, example.belemnite, example.breccia,
  example.lense, example.liquefaction, fossil.example, proxy.example,
  proxy.example.litho}
  \item{\strong{Magnetostratigraphical data}}{chron.example}
  \item{\strong{Litholog exportation script}}{log.loop.tex}
  \item{\strong{Oriented data}}{zeq_example}
}
}
